backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
supabase==2.9.0
google-cloud-speech==2.28.0
google-cloud-translate==3.17.0
google-generativeai==0.8.3
sentence-transformers==3.2.1
openai==1.54.4
python-dotenv==1.0.1
pydantic==2.9.2
